# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant PIM API
  version: 2.4.0
  description: |
    API para gestión de información de productos multi-tenant.

    ## Arquitectura Modular

    El sistema está organizado en módulos especializados:
    - **Health**: Estado del sistema
    - **Business Types**: Tipos de negocio
    - **Brands**: Gestión de marcas
    - **Products**: Gestión de productos y variantes
    - **Variants**: Gestión de variantes de productos (HITO 2.1)
    - **Categories**: Gestión de categorías
    - **Attributes**: Gestión de atributos
    - **Category-Attributes**: Relación categoría-atributo
    - **Global Catalog**: Catálogo global de productos con enrichment y quality score
    - **Global Catalog Public**: Endpoints públicos del catálogo global (sin auth)
    - **Product Requests**: Solicitudes de productos por tenants
    - **Quickstart**: Configuración rápida e importación desde global catalog
    - **Import**: Importación masiva con variantes (HITO 2.1)

    ## Actualizado v2.3.0

    - Global Catalog: CRUD completo, búsqueda por EAN, enrichment queue
    - Product Requests: Solicitudes de tenants con flujo approve/reject/fulfill
    - Quickstart: Import desde global catalog, import masivo por business type
    - Public endpoints: Búsqueda, sugerencias, facets sin autenticación
    - Variantes de productos: Product -> Variants con SKU único
    - Templates Quickstart: Configuración rápida por tipo de negocio
...
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: ID del tenant para operaciones multi-tenant
    userRole:
      type: apiKey
      in: header
      name: X-User-Role
      description: |
        Rol del usuario para autorización:
        - `super_admin`: Acceso completo
        - `marketplace_admin`: Gestión de marketplace
        - `tenant_admin`: Administración del tenant
        - `tenant_user`: Usuario del tenant
  parameters:
    PageParam:
      name: page
      in: query
      description: "Número de página (default: 1)"
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: "Tamaño de página (default: 10, max: 100)"
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortByParam:
      name: sort_by
      in: query
      description: "Campo de ordenamiento (default: created_at)"
      required: false
      schema:
        type: string
        default: "created_at"
    SortDirParam:
      name: sort_dir
      in: query
      description: "Dirección de ordenamiento (asc o desc, default: desc)"
      required: false
      schema:
        type: string
        enum: ["asc", "desc"]
        default: "desc"
...
paths:
...
  /products:
    get:
      summary: Listar productos con filtros
      tags: [Products]
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirParam'
        - name: name
          in: query
          description: Filtrar por nombre
          schema:
            type: string
        - name: sku
          in: query
          description: Filtrar por SKU
          schema:
            type: string
        - name: status
          in: query
          description: Filtrar por estado
          schema:
            type: string
            enum: ["active", "inactive", "discontinued"]
        - name: category_id
          in: query
          description: Filtrar por ID de categoría
          schema:
            type: string
            format: uuid
        - name: brand_id
          in: query
          description: Filtrar por ID de marca
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista paginada de productos
        '400':
          description: Parámetros inválidos
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

    post:
      summary: Crear nuevo producto
      tags: [Products]
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, sku]
              properties:
                name:
                  type: string
                  example: "Smartphone Galaxy S23"
                description:
                  type: string
                  example: "Smartphone de última generación"
                sku:
                  type: string
                  example: "SAM-GAL-S23-001"
                category_id:
                  type: string
                  format: uuid
                brand_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Producto creado exitosamente
        '400':
          description: Datos de entrada inválidos
        '409':
          description: SKU ya existe
        '500':
          description: Error interno del servidor
...
  /products/import:
    post:
      summary: Importar productos con variantes
      tags: [Import]
      description: |
        Importa productos con sus variantes desde CSV o JSON.

        El parser detecta automáticamente:
        - Columnas estándar (name, sku, price, category, brand)
        - Atributos dinámicos (potencia, voltaje, color, etc.)
        - Agrupa filas por product_name para crear variantes

        Formato CSV esperado:
        ```
        product_name,variant_sku,category,brand,potencia,voltaje,price
        Taladro Bosch,TALADRO-750W,Herramientas,Bosch,750W,220V,45000
        Taladro Bosch,TALADRO-1000W,Herramientas,Bosch,1000W,220V,65000
        ```
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo CSV o JSON
                create_categories:
                  type: boolean
                  default: true
                  description: Crear categorías automáticamente si no existen
                create_brands:
                  type: boolean
                  default: true
                  description: Crear marcas automáticamente si no existen
          application/json:
            schema:
              type: object
              required:
                - products
              properties:
                products:
                  type: array
                  items:
                    type: object
                    properties:
                      product_name:
                        type: string
                      category_name:
                        type: string
                      brand_name:
                        type: string
                      variants:
                        type: array
                        items:
                          type: object
                          properties:
                            sku:
                              type: string
                            attributes:
                              type: object
                            price:
                              type: number
                create_categories:
                  type: boolean
                  default: true
                create_brands:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Resultado de la importación
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  summary:
                    type: object
                    properties:
                      products_created:
                        type: integer
                      variants_created:
                        type: integer
                      categories_created:
                        type: integer
                      brands_created:
                        type: integer
                      products_skipped:
                        type: integer
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            row:
                              type: integer
                            error:
                              type: string
        '400':
          description: Archivo inválido o formato incorrecto
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

  # ==========================================
  # QUICKSTART TEMPLATES ENDPOINTS (HITO 2)
  # ==========================================
...